pyproject.toml
[tool.poetry]
name = "python-rule-engine-benchmark"
version = "0.1.0"
description = ""
authors = ["MrLYC <imyikong@gmail.com>"]
readme = "README.md"
packages = [{include = "python_rule_engine_benchmark"}]

[tool.poetry.dependencies]
python = "^3.11"
cdmn = "^2.0.2"
zen-engine = "^0.6.0"
pydmnrules = "^1.4.4"
rule-engine = "^3.5.1"
business-rules = "^1.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-lazy-fixture = "^0.6.3"
pytest-benchmark = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
